feat: implement Read Aloud engine with Puter AI and browser TTS support via new composables
