+++ docs/.vitepress/theme/components/ReadAloud.vue
@@ -1,414 +1,26 @@
 <script setup lang="ts">
-import { ref, onMounted, onUnmounted, watch } from 'vue'
-import { useRoute } from 'vitepress'
+import { useReadAloudEngine } from '../composables/useReadAloudEngine'
 
-const route = useRoute()
-
-const isSupported = ref(true)
-const isPlaying = ref(false)
-const isPaused = ref(false)
-const isLoadingModel = ref(false)
-const loadingProgress = ref('')
-const currentRate = ref(1.0)
-const rates = [0.8, 1.0, 1.25, 1.5, 2.0]
-
-// Engine selection
-const ttsMode = ref<'puter' | 'browser'>('puter')
-
-// Puter OpenAI voices
-const puterVoices = [
-  { id: 'nova', name: 'Nova (Female - Natural)', provider: 'openai' },
-  { id: 'alloy', name: 'Alloy (Neutral - Clear)', provider: 'openai' },
-  { id: 'echo', name: 'Echo (Male - Warm)', provider: 'openai' },
-  { id: 'fable', name: 'Fable (Expressive - Story)', provider: 'openai' },
-  { id: 'onyx', name: 'Onyx (Male - Deep)', provider: 'openai' },
-  { id: 'shimmer', name: 'Shimmer (Female - Soft)', provider: 'openai' },
-]
-const selectedPuterVoice = ref('nova')
-
-// Browser voices
-const browserVoices = ref<SpeechSynthesisVoice[]>([])
-const selectedBrowserVoiceIndex = ref(0)
-
-const chunks = ref<string[]>([])
-const currentChunkIndex = ref(0)
-let currentAudio: HTMLAudioElement | null = null
-let currentUtterance: SpeechSynthesisUtterance | null = null
-
-// Audio cache and background prefetch maps
-const audioCache = new Map<string, HTMLAudioElement>()
-const prefetchMap = new Map<string, Promise<HTMLAudioElement | null>>()
-
-const clearAudioCache = () => {
-  audioCache.forEach(audio => {
-    audio.pause()
-    audio.src = ''
-  })
-  audioCache.clear()
-  prefetchMap.clear()
-}
-
-const stopSpeech = () => {
-  isPlaying.value = false
-  isPaused.value = false
-  isLoadingModel.value = false
-  loadingProgress.value = ''
-  currentChunkIndex.value = 0
-  chunks.value = []
-  
-  if (currentAudio) {
-    currentAudio.pause()
-    currentAudio = null
-  }
-
-  if (typeof window !== 'undefined' && 'speechSynthesis' in window) {
-    window.speechSynthesis.cancel()
-  }
-  currentUtterance = null
-}
-
-const getArticleText = (): string => {
-  if (typeof document === 'undefined') return ''
-  const docElement = document.querySelector('.vp-doc')
-  if (!docElement) return ''
-
-  const clone = docElement.cloneNode(true) as HTMLElement
-  
-  // Remove non-text UI elements
-  clone.querySelectorAll('.line-numbers-wrapper, .copy, .header-anchor, script, style, .read-aloud-container, nav, footer, .VPNav, .VPFooter, button, img, svg, .vp-code, pre, code').forEach(el => el.remove())
-  
-  let text = clone.innerText || clone.textContent || ''
-  
-  // Normalize text for natural speech pacing
-  text = text
-    .replace(/#{1,6}\s+/g, '. ')
-    .replace(/^[\s]*[-*+]\s+/gm, '. ')
-    .replace(/\n+/g, '. ')
-    .replace(/\s+/g, ' ')
-    .replace(/\b(e\.g\.|i\.e\.|vs\.|etc\.)/gi, match => match.replace(/\./g, ''))
-    .replace(/([.!?])\s*([A-Z])/g, '$1 $2')
-    .replace(/`([^`]+)`/g, '$1')
-    .replace(/\[([^\]]+)\]\([^)]+\)/g, '$1')
-    .trim()
-    
-  return text
-}
-
-const splitIntoChunks = (text: string): string[] => {
-  const rawSentences = text.match(/[^.!?\n]+[.!?\n]+/g) || [text]
-  const result: string[] = []
-  let current = ''
-
-  for (const sentence of rawSentences) {
-    const trimmed = sentence.trim()
-    if (!trimmed) continue
-    if ((current + ' ' + trimmed).length > 250) {
-      if (current) result.push(current.trim())
-      current = trimmed
-    } else {
-      current = current ? current + ' ' + trimmed : trimmed
-    }
-  }
-  if (current.trim()) result.push(current.trim())
-  return result.length > 0 ? result : [text]
-}
-
-const loadBrowserVoices = () => {
-  if (typeof window !== 'undefined' && 'speechSynthesis' in window) {
-    const voices = window.speechSynthesis.getVoices()
-    if (!voices || voices.length === 0) return
-    
-    const scoreVoice = (v: SpeechSynthesisVoice) => {
-      let score = 0
-      const name = v.name.toLowerCase()
-      if (name.includes('google') && name.includes('us english')) score += 1000
-      if (name.includes('microsoft') && (name.includes('jenny') || name.includes('guy') || name.includes('aria'))) score += 1000
-      if (name.includes('samantha') || name.includes('alex') || name.includes('daniel')) score += 900
-      if (name.includes('google') || name.includes('microsoft') || name.includes('natural')) score += 500
-      if (v.lang === 'en-US') score += 100
-      else if (v.lang.startsWith('en')) score += 50
-      if (v.default) score += 200
-      return score
-    }
-    
-    const sortedVoices = [...voices].sort((a, b) => scoreVoice(b) - scoreVoice(a))
-    const englishVoices = sortedVoices.filter(v => v.lang.startsWith('en'))
-    browserVoices.value = englishVoices.length > 0 ? englishVoices : sortedVoices
-  }
-}
-
-// Prefetch audio chunk in background and store in cache
-const fetchChunkAudio = (chunkIndex: number): Promise<HTMLAudioElement | null> => {
-  if (chunkIndex < 0 || chunkIndex >= chunks.value.length) {
-    return Promise.resolve(null)
-  }
-
-  const cacheKey = `${selectedPuterVoice.value}:${chunkIndex}:${chunks.value[chunkIndex]}`
-
-  if (audioCache.has(cacheKey)) {
-    return Promise.resolve(audioCache.get(cacheKey)!)
-  }
-
-  if (prefetchMap.has(cacheKey)) {
-    return prefetchMap.get(cacheKey)!
-  }
-
-  const promise = (async () => {
-    try {
-      const puterModule = await import('@heyputer/puter.js')
-      const puter = puterModule.default || puterModule.puter || puterModule
-
-      const audio = await puter.ai.txt2speech(chunks.value[chunkIndex], {
-        provider: 'openai',
-        model: 'tts-1',
-        voice: selectedPuterVoice.value,
-      })
-
-      audioCache.set(cacheKey, audio)
-      prefetchMap.delete(cacheKey)
-      return audio
-    } catch (err) {
-      console.warn(`Failed to prefetch Puter AI chunk ${chunkIndex}:`, err)
-      prefetchMap.delete(cacheKey)
-      return null
-    }
-  })()
-
-  prefetchMap.set(cacheKey, promise)
-  return promise
-}
-
-const speakCurrentPuterChunk = async () => {
-  if (currentChunkIndex.value >= chunks.value.length) {
-    isPlaying.value = false
-    isPaused.value = false
-    isLoadingModel.value = false
-    loadingProgress.value = ''
-    currentChunkIndex.value = 0
-    return
-  }
-
-  const idx = currentChunkIndex.value
-  const cacheKey = `${selectedPuterVoice.value}:${idx}:${chunks.value[idx]}`
-
-  // Show loading indicator only if chunk is not yet cached
-  if (!audioCache.has(cacheKey)) {
-    isLoadingModel.value = true
-    loadingProgress.value = `Loading AI audio (${idx + 1}/${chunks.value.length})...`
-  }
-
-  const audio = await fetchChunkAudio(idx)
-
-  if (!audio || !isPlaying.value) {
-    isLoadingModel.value = false
-    loadingProgress.value = ''
-    if (!audio) {
-      // Fallback to browser if Puter API fails
-      ttsMode.value = 'browser'
-      speakCurrentBrowserChunk()
-    }
-    return
-  }
-
-  isLoadingModel.value = false
-  loadingProgress.value = ''
-
-  if (currentAudio && currentAudio !== audio) {
-    currentAudio.pause()
-  }
-
-  currentAudio = audio
-  currentAudio.currentTime = 0
-  currentAudio.playbackRate = currentRate.value
-
-  // Trigger background pre-fetch for the NEXT 2 chunks so playback never waits!
-  fetchChunkAudio(idx + 1)
-  fetchChunkAudio(idx + 2)
-
-  audio.onended = () => {
-    if (!isPlaying.value || isPaused.value) return
-    currentChunkIndex.value++
-    if (currentChunkIndex.value < chunks.value.length) {
-      speakCurrentPuterChunk()
-    } else {
-      isPlaying.value = false
-      isPaused.value = false
-      currentChunkIndex.value = 0
-    }
-  }
-
-  audio.onerror = (e: any) => {
-    console.warn('Puter audio chunk error, skipping to next:', e)
-    currentChunkIndex.value++
-    if (isPlaying.value && !isPaused.value && currentChunkIndex.value < chunks.value.length) {
-      speakCurrentPuterChunk()
-    } else {
-      isPlaying.value = false
-      isPaused.value = false
-    }
-  }
-
-  await audio.play().catch(() => {})
-}
-
-const speakCurrentBrowserChunk = () => {
-  if (typeof window === 'undefined' || !('speechSynthesis' in window)) return
-  if (currentChunkIndex.value >= chunks.value.length) {
-    isPlaying.value = false
-    isPaused.value = false
-    currentChunkIndex.value = 0
-    return
-  }
-
-  window.speechSynthesis.cancel()
-
-  const text = chunks.value[currentChunkIndex.value]
-  const utterance = new SpeechSynthesisUtterance(text)
-  const voice = browserVoices.value[selectedBrowserVoiceIndex.value]
-  
-  if (voice) {
-    utterance.voice = voice
-    utterance.lang = voice.lang
-  }
-  
-  utterance.rate = currentRate.value
-  utterance.pitch = 1.0
-  utterance.volume = 1.0
-  
-  utterance.onstart = () => {
-    isPlaying.value = true
-    isPaused.value = false
-  }
-  
-  utterance.onend = () => {
-    if (!isPlaying.value || isPaused.value) return
-    currentChunkIndex.value++
-    if (currentChunkIndex.value < chunks.value.length) {
-      speakCurrentBrowserChunk()
-    } else {
-      isPlaying.value = false
-      isPaused.value = false
-      currentChunkIndex.value = 0
-    }
-  }
-  
-  utterance.onerror = (event) => {
-    if (event.error === 'canceled' || event.error === 'interrupted') return
-    console.error('Speech synthesis error:', event.error)
-    currentChunkIndex.value++
-    if (isPlaying.value && !isPaused.value && currentChunkIndex.value < chunks.value.length) {
-      speakCurrentBrowserChunk()
-    } else {
-      isPlaying.value = false
-      isPaused.value = false
-    }
-  }
-
-  currentUtterance = utterance
-  window.speechSynthesis.speak(utterance)
-}
-
-const togglePlayPause = async () => {
-  if (isPlaying.value && !isPaused.value) {
-    // Pause
-    isPaused.value = true
-    if (ttsMode.value === 'browser') {
-      if (typeof window !== 'undefined' && 'speechSynthesis' in window) {
-        window.speechSynthesis.cancel()
-      }
-    } else {
-      if (currentAudio) {
-        currentAudio.pause()
-      }
-    }
-    return
-  }
-
-  if (isPaused.value) {
-    // Resume
-    isPaused.value = false
-    isPlaying.value = true
-    if (ttsMode.value === 'browser') {
-      speakCurrentBrowserChunk()
-    } else {
-      if (currentAudio && currentAudio.paused) {
-        currentAudio.play().catch(() => {
-          speakCurrentPuterChunk()
-        })
-      } else {
-        speakCurrentPuterChunk()
-      }
-    }
-    return
-  }
-
-  // Start new speech
-  stopSpeech()
-  const fullText = getArticleText()
-  if (!fullText) return
-
-  chunks.value = splitIntoChunks(fullText)
-  currentChunkIndex.value = 0
-  isPlaying.value = true
-  isPaused.value = false
-
-  if (ttsMode.value === 'puter') {
-    // Start pre-fetching first 2 chunks immediately
-    fetchChunkAudio(0)
-    fetchChunkAudio(1)
-    await speakCurrentPuterChunk()
-  } else {
-    speakCurrentBrowserChunk()
-  }
-}
-
-const setRate = (rate: number) => {
-  currentRate.value = rate
-  if (ttsMode.value === 'puter') {
-    if (currentAudio) {
-      currentAudio.playbackRate = rate
-    }
-  } else {
-    if (isPlaying.value && !isPaused.value) {
-      speakCurrentBrowserChunk()
-    }
-  }
-}
-
-const onEngineChange = () => {
-  if (isPlaying.value || isPaused.value) {
-    stopSpeech()
-  }
-}
-
-const onVoiceChange = () => {
-  clearAudioCache()
-  if (isPlaying.value || isPaused.value) {
-    stopSpeech()
-    setTimeout(() => togglePlayPause(), 50)
-  }
-}
-
-onMounted(() => {
-  if (typeof window !== 'undefined') {
-    if (!('speechSynthesis' in window)) {
-      isSupported.value = false
-    } else {
-      loadBrowserVoices()
-      window.speechSynthesis.onvoiceschanged = loadBrowserVoices
-    }
-  }
-})
-
-watch(() => route.path, () => {
-  stopSpeech()
-  clearAudioCache()
-})
-
-onUnmounted(() => {
-  stopSpeech()
-  clearAudioCache()
-})
+const {
+  isSupported,
+  ttsMode,
+  isPlaying,
+  isPaused,
+  isLoadingModel,
+  loadingProgress,
+  currentRate,
+  rates,
+  puterVoices,
+  selectedPuterVoice,
+  browserVoices,
+  selectedBrowserVoiceIndex,
+  isBrowserSupported,
+  togglePlayPause,
+  stopSpeech,
+  setRate,
+  onEngineChange,
+  onVoiceChange,
+} = useReadAloudEngine()
 </script>
 
 <template>
@@ -457,7 +69,7 @@ onUnmounted(() => {
           title="Select TTS Engine"
         >
           <option value="puter">Puter AI (Natural OpenAI)</option>
-          <option value="browser">Browser Native (Offline)</option>
+          <option v-if="isBrowserSupported" value="browser">Browser Native (Offline)</option>
         </select>
 
         <!-- Voice Selector for Puter AI -->
@@ -479,7 +91,7 @@ onUnmounted(() => {
 
         <!-- Voice Selector for Browser Native -->
         <select 
-          v-else
+          v-else-if="ttsMode === 'browser' && isBrowserSupported"
           v-model="selectedBrowserVoiceIndex" 
           @change="onVoiceChange"
           class="voice-select"
